Migrated header component.
